thread events

+++ src/controllers/core/Thread.ts
@@ -1,47 +1,60 @@
 import {ThrottleCallback, TransferData} from '../../types/core/Threads'
-import ThreadInterface, {Mode, State} from '../../types/core/Thread'
+import ThreadInterface, {Mode, State, Event} from '../../types/core/Thread'
 import {HybridExecutor} from '../../types/core/Executor'
 import {Task} from '../../types/partials/TaskPool'
 
 
 export default class Thread implements ThreadInterface {
-    readonly #mode: Mode = Mode.PARALLEL
-
     #executor: HybridExecutor
 
     #state: State = State.IDLE
 
-    constructor(Executor: new() => HybridExecutor, mode: Mode) {
+    #events: Map<Event, Function[]> = new Map()
+
+    constructor(Executor: new() => HybridExecutor) {
         this.#executor = new Executor()
-        this.#mode = mode
     }
 
-    async execute(data: TransferData): Promise<any[]> {
+    async execute(data: TransferData, mode: Mode = Mode.SEQUENTIAL): Promise<any[]> {
+        if (this.#state !== State.IDLE) throw 'Thread is already running (Internal Controller Error)'
+
         this.#state = State.RUNNING
+
+        this.#emit(Event.PROGRESS, this)
 
         const responses: any[] = data?.responses ?? []
 
-        const tasks = data.pool.pool
+        const tasks: Task[] = data.pool.pool
 
         while (data.pool.length && this.#state === State.RUNNING) {
             // Get the next task & remove it from the pool
             const task: Task = tasks.splice(0, 1).at(0)!
 
             // Sequential mode: set the message to the response of the previous task (if exists)
-            if (this.#mode === Mode.SEQUENTIAL) task.message = task?.message ?? responses.at(-1)
+            if (mode === Mode.SEQUENTIAL) task.message = task?.message ?? responses.at(-1)
 
             // If throttle is set, wait for its completion. If it fails, terminate the thread
-            if (data.throttle) await this.#waitForThrottleSuccess(data.throttle).catch(() => {
+            if (data.throttle) await this.#waitForThrottleSuccess(data.throttle).catch()
+
+            //@TODO: fix TS2367
+            //@ts-ignore
+            if(this.#state === State.INTERRUPTED) {
                 this.#state = State.IDLE
                 this.#executor.terminate()
-            })
+
+                this.#emit(Event.ERROR, this)
+
+                break
+            }
 
             // Run the task and get the response
             const response = await this.#executor.run(task.method, task.message)
 
             // Check if the response is an error and gracefully terminate the thread
-            if (response.error) {
+            if (response?.error) {
                 console.error(response.error)
+                this.#emit(Event.ERROR, this)
+
                 break
             }
 
@@ -49,45 +62,50 @@ export default class Thread implements ThreadInterface {
             responses[task.index!] = response
 
             // Execute the step callback
-            const progress: number = 100 * responses.filter((response) => response !== undefined).length / data.poolSize
-            data.step?.(responses[task.index!], progress)
-
+            const progress: number = responses.filter((response) => response !== undefined).length / data.poolSize
+            data.step?.(responses[task.index!], progress * 100)
         }
 
         this.#state = State.IDLE
         this.#executor.terminate()
 
+        this.#emit(Event.COMPLETE, this)
+
         return responses
     }
 
     terminate(): void {
-        this.#state = State.IDLE
+        this.#state = State.INTERRUPTED
     }
 
-    #waitForThrottleSuccess = async (throttle: ThrottleCallback): Promise<void> => {
-        return await new Promise<void>(async (resolve) => {
+    on(event: Event, callback: (data: any) => void): void {
+        if (!this.#events.has(event)) this.#events.set(event, [])
+
+        this.#events.get(event)!.push(callback)
+    }
+
+    #emit(event: Event, data: any): void {
+        if (!this.#events.has(event)) return
+
+        this.#events.get(event)!.forEach((callback) => callback(data))
+    }
+
+    async #waitForThrottleSuccess(throttle: ThrottleCallback): Promise<void> {
+        while (true) {
+            if (this.#state === State.INTERRUPTED) break
+
             try {
-                if (await throttle()) return resolve()
+                if (await throttle()) break
             }
             catch (e) {
-                throw new Error(`Throttle error: ${e}`)
+                this.#state = State.INTERRUPTED
+
+                throw `Throttle error: ${e}`
             }
 
-            const interval: NodeJS.Timeout | number = setInterval(async () => {
-                try {
-                    if (await throttle()) {
-                        clearInterval(interval)
-                        resolve()
-                    }
-                }
-                catch (e) {
-                    clearInterval(interval)
-                    throw new Error(`Throttle error: ${e}`)
-                }
-            }, 50)
-        })
+            await new Promise(requestAnimationFrame)
+        }
     }
-
 
     get state(): State {
         return this.#state
